Define base class for delivery service used for embedapi v2 and v3
